+++ CHANGELOG.md
@@ -90,3 +90,4 @@ For more information, see the [Advanced Functionalities - Advanced control over 
 - Synchronous classes are no longer available, all async classes have been renamed.
 - .NET 6 is no longer supported.
 - NotoriousTest is now in .net standard 2.1, web support has been moved to NotoriousTest.Web
+- XUnit has been updated to the latest version (xunit.v3).